airflow/requirements.txt
dbt-core
dbt-postgres
pandas
investpy
psycopg2-binary
pandasql
python-dotenv
supabase
SQLAlchemy
apache-airflow-providers-postgres
apache-airflow-providers-apache-spark
pyarrow
pyspark
celery
kombu
tornado
werkzeug
wtforms
bs4